check-ceph: fix error with '--ignore-flags'
This patch fixes an error that occurs if '--ignore-flags' is set but no
flag is reported set by ceph health.

Method chaining in strip_warns fails if the ceph health output does not
match the gsub! patterns "/HEALTH_WARN\ /" and "/\ ?flag\(s\) set/"
(because gsub! returns nil if nothing matches).